Added new admin layout,refactoring code and new features
Admin: Added a new admin layout. It has the tabs I want to
create, but only works the one to delete backups. Must be
improved, but it's a first approach, and it deletes the
checked backups.

Modified how the variables are saved. Now it's saved all the
user's preferences in a uniq preference

Modifications when saving secrets and its backups. Now we can
have a "log" for the secrets, to know who and what has done.

Added favicons.

Closes #34 #36 #38
References #53